refactor(analysis): lift face detection out of the scoring module
scoring.py was 845 lines carrying four unrelated jobs: face detection,
motion/duration scoring, segment generation, and SceneScorer. Face
detection is the one that comes away cleanly -- it is the only part with
platform-specific backends, its own availability probe, and module-level
cached state, and no module in src/ imported it. It was internal to
SceneScorer and to tests.

scoring.py 845 -> 684, face_scoring.py at 186.

**The imports in scoring.py are use, not re-export.** SceneScorer calls
check_vision_available, init_vision_detector and init_opencv_cascade in its
constructor and compute_face_score when scoring a scene, so the names have
to resolve in scoring's namespace. That also means the twelve existing
`patch("immich_memories.analysis.scoring.check_vision_available")` targets
keep working untouched: `from x import y` binds y in the importing module,
so scoring's namespace is still where SceneScorer looks the name up.
Patching face_scoring instead would silently fail to affect the scorer,
which is why those targets are deliberately left alone.

Tests that import a moved symbol directly, or reach into `_use_vision`, or
patch `platform`/`cv2` for a moved function, are repointed at
face_scoring -- scoped to the three test classes that cover moved
functions, so the SceneScorer tests that patch `scoring.cv2.VideoCapture`
stay as they are.

`platform` is dropped from scoring.py's imports; it went with the face
code. face_scoring joins scoring in the mypy override for modules using
OpenCV APIs with incomplete stubs -- `cv2.data` has no annotation.

No behaviour change: the only added lines in scoring.py are the docstring
and the import block. No logic moved into or out of a branch.

Noted while mapping this, not fixed here: `_check_vision_available` is
implemented three times -- here, in analysis/apple_vision.py and in
processing/transforms_smart_crop.py. Consolidating them is a behaviour
question, not a file split.

Refs #245
